Move GetTouristByTourId to tag joinings
